+++ src/realm/sync/network/websocket.cpp
@@ -770,6 +770,8 @@ private:
             ec = Error::bad_response_2xx_successful;
         else if (status_code == 301)
             ec = Error::bad_response_301_moved_permanently;
+        else if (status_code == 308)
+            ec = Error::bad_response_308_permanent_redirect;
         else if (status_code >= 300 && status_code < 400)
             ec = Error::bad_response_3xx_redirection;
         else if (status_code == 401)
@@ -1047,6 +1049,8 @@ const char* get_error_message(Error error_code)
             return "Bad WebSocket response 3xx redirection";
         case Error::bad_response_301_moved_permanently:
             return "Bad WebSocket response 301 moved permanently";
+        case Error::bad_response_308_permanent_redirect:
+            return "Bad WebSocket response 308 permanent redirect";
         case Error::bad_response_4xx_client_errors:
             return "Bad WebSocket response 4xx client errors";
         case Error::bad_response_401_unauthorized:
@@ -1105,36 +1109,10 @@ class CloseStatusErrorCategory : public std::error_category {
     {
         // Converts an error_code to one of the pre-defined status codes in
         // https://tools.ietf.org/html/rfc6455#section-7.4.1
-        switch (error_code) {
-            case 1000:
-                return "normal closure";
-            case 1001:
-                return "endpoint going away";
-            case 1002:
-                return "protocol error";
-            case 1003:
-                return "invalid data type";
-            case 1004:
-                return "reserved";
-            case 1005:
-                return "no status code present";
-            case 1006:
-                return "no close control frame sent";
-            case 1007:
-                return "message data type mis-match";
-            case 1008:
-                return "policy violation";
-            case 1009:
-                return "message too big";
-            case 1010:
-                return "missing extension";
-            case 1011:
-                return "unexpected error";
-            case 1015:
-                return "TLS handshake failure";
-            default:
-                return "unknown error";
-        };
+        if (error_code == 1000 || error_code == 0) {
+            return ErrorCodes::error_string(ErrorCodes::OK);
+        }
+        return ErrorCodes::error_string(static_cast<ErrorCodes::Error>(error_code));
     }
 };
 
@@ -1254,15 +1232,20 @@ util::Optional<HTTPResponse> websocket::make_http_response(const HTTPRequest& re
     return do_make_http_response(request, sec_websocket_protocol, ec);
 }
 
-const std::error_category& websocket::error_category() noexcept
-{
-    return g_error_category;
-}
-
 const std::error_category& websocket::websocket_close_status_category() noexcept
 {
     static const CloseStatusErrorCategory category = {};
     return category;
+}
+
+std::error_code websocket::make_error_code(ErrorCodes::Error error) noexcept
+{
+    return std::error_code{error, realm::sync::websocket::websocket_close_status_category()};
+}
+
+const std::error_category& websocket::error_category() noexcept
+{
+    return g_error_category;
 }
 
 std::error_code websocket::make_error_code(Error error_code) noexcept